backend/services/django/requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
coreschema==0.0.4
Django==5.2
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
whitenoise==6.9.0
